# Services/rooting/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Graph API",
    "description": "This API can create chart from dataSets"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/graphs/": {
      "post": {
        "tags": [
          "Graphs"
        ],
        "description": "Creation of a graph",
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "description": "Type of graph created : histogramme,connectedLine,doughnut,circlePoint or bezierCurve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Return the id of the graph created"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "501": {
            "description": "Return the error message"
          }
        }
      }
    },
    "/graphs/{id}/": {
      "put": {
        "tags": [
          "Graphs"
        ],
        "description": "Modify a specific graph",
        "parameters": [
          {
            "name": "id",
            "in": "param",
            "required": true,
            "description": "The id of the graph",
            "type": "integer"
          },
          {
            "name": "information to modify",
            "in": "body",
            "description": "Element of the graph to modify",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/graph"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "MODIFY"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "500": {
            "description": "Return the error message"
          }
        }
      },
      "delete": {
        "tags": [
          "Graphs"
        ],
        "description": "Suppression of a specific graph",
        "parameters": [
          {
            "name": "id",
            "in": "param",
            "required": true,
            "description": "The id of the graph",
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "DELETED"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "501": {
            "description": "Return the error message"
          }
        }
      }
    },
    "/graphs/{graphId}/dataSet/{dataSetId}": {
      "post": {
        "tags": [
          "Link"
        ],
        "description": "Link a graph and a dataSet",
        "parameters": [
          {
            "name": "graphId",
            "in": "param",
            "required": true,
            "description": "Id of the graph to link",
            "type": "integer"
          },
          {
            "name": "dataSetId",
            "in": "param",
            "required": true,
            "description": "Id of the dataSet to link",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "ADDED"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "500": {
            "description": "Return the error message"
          }
        }
      },
      "delete": {
        "tags": [
          "Link"
        ],
        "description": "Delete the link between a graph and a dataSet",
        "parameters": [
          {
            "name": "graphId",
            "in": "param",
            "required": true,
            "description": "Id of the graph to unlink",
            "type": "integer"
          },
          {
            "name": "dataSetId",
            "in": "param",
            "required": true,
            "description": "Id of the dataSet to unlink",
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "description": "REMOVED"
          },
          "400": {
            "description": "Return the validation error array from express validator"
          },
          "500": {
            "description": "Return the error message"
          }
        }
      }
    },
...
    "/graphs/{graphId}": {
      "get": {
        "tags": [
          "Graphs"
        ],
        "description": "Get all information about a graph",
        "parameters": [
          {
            "name": "graphId",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "param",
            "description": "The id of the graph",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Return the graph charas",
            "schema": {
              "$ref": "#/definitions/graph"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "graph": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Super graph !"
        },
        "description": {
          "type": "string",
          "example": "A super graph"
        },
        "type": {
          "type": "string",
          "example": "circlePoint"
        },
        "showX": {
          "type": "boolean",
          "example": true
        },
        "showGrid": {
          "type": "boolean",
          "example": true
        },
        "xBounds": {
          "type": "object",
          "properties": {
            "lowerBound": {
              "type": "number",
              "example": 10
            },
            "upperBound": {
              "type": "number",
              "example": 15.54
            }
          }
        },
        "xLegend": {
          "type": "string",
          "example": "My x axe"
        },
        "graphLegend": {
          "type": "string",
          "example": "My legend"
        },
        "backgroundColor": {
          "type": "object",
          "properties": {
            "r": {
              "type": "number",
              "example": 12
            },
            "g": {
              "type": "number",
              "example": 13
            },
            "b": {
              "type": "number",
              "example": 14
            },
            "a": {
              "type": "number",
              "example": 15
            }
          }
        },
        "gridColor": {
          "type": "object",
          "properties": {
            "r": {
              "type": "number",
              "example": 12
            },
            "g": {
              "type": "number",
              "example": 13
            },
            "b": {
              "type": "number",
              "example": 14
            },
            "a": {
              "type": "number",
              "example": 15
            }
          }
        }
      },
      "xml": {
        "name": "graph"
      }
    },
...
  }
}